Fix issue with disappearing search results panel
There was previously an issue (undocumented) where after adding a song
from the search results list, the search box would seem to stop to function
because it the search results wouldn't reappear/populate. It was an issue
where we were hiding the div (with `slideUp()`) and never showing it on
re-render. I fixed that with a call to `slideDown()` on render which shows
the div again.

Doing this I also refactored clearing the search box and related
requests/javascript data into a function called `clearSearch()`.